+++ utils/fgo-spreadsheet/makeMaterialDataRow.js
@@ -1,0 +1,10 @@
+export function makeMaterialDataRow(material) {
+  const row = [
+    material.id,
+    material.name,
+    material.background,
+    `=IMAGE("${material.icon}")`
+  ];
+
+  return row.join("\t");
+}
